Fix a user facing errno on the last passwd or group entry.
Original bug was observed in ghc testsuite:
    https://github.com/ghc/packages-unix/blob/master/System/Posix/User.hsc#L374
which checks error condition as a 'NULL' result and 'errno != 0'.

The problem is in asymmetry of 'fgetpwent_r()' used
to implement 'getpwent()'. In case of end of file
'fgetpwent_r()' returns 'NULL' and 'errno == ENOENT'
and nsscache propagates that errno value to glibc.

Signed-off-by: Sergei Trofimovich <slyfox@gentoo.org>