Stop installing multiple copies of the shared library.
This is never linked into programs by ld, so the symlink redirections from .so
and soname are not required.  Moreover, the Debian package was ending up with
three copies of the same file (rather than one file and two symlinks).

The only copy will be installed as libnss_cache.so.2, to meet the API for NSS
which expects to load libnss_<service>.so.<x>, where x is 2 for glibc 2.1 and
later (see nsswitch.conf(5)).